Bugfix: when using new twitter accounts, you could receive the following error: Error 404 - this page does not exists (or server 500 error in production) This is caused by using a to small storage field (int) for twitter_users.twitter_id, changed the field to a bigint.

+++ src/Backend/Modules/Twitter/Installer/Data/install.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS `twitter_tweets` (
 
 CREATE TABLE IF NOT EXISTS `twitter_users` (
  `id` int(11) NOT NULL AUTO_INCREMENT,
- `twitter_id` int(11) DEFAULT NULL,
+ `twitter_id` bigint(20) unsigned DEFAULT NULL,
  `username` varchar(50) NOT NULL,
  `name` varchar(255) DEFAULT NULL,
  `location` varchar(255) DEFAULT NULL,
